backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.123.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]

[tool.uv.scripts]
dev = "uvicorn main:app --reload"
